api/requirements.txt
anthropic==0.29.0
Flask==3.0.3
Flask_Cors==3.0.10
Flask_SocketIO==5.3.6
groq==0.9.0
openai==1.35.3
Requests==2.32.3
azure.storage.blob==12.20.0